update to exceptions and rearrangement of files to ensure runtest.sh works
