--- src/main/java/duke/exceptions/DukeException.java
@@ -1,7 +1,0 @@
-package duke.exceptions;
-
-public class DukeException extends Exception{
-    public DukeException(String errorMessage) {  
-        super(errorMessage);  
-    }  
-}
