+++ src/main/java/exceptions/InvalidDeadlineException.java
@@ -1,0 +1,5 @@
+package exceptions;
+
+public class InvalidDeadlineException extends Exception {
+    // no further code
+}
