+++ src/main/java/exceptions/InvalidEventException.java
@@ -1,0 +1,5 @@
+package exceptions;
+
+public class InvalidEventException extends Exception{
+     // no further code
+}
