+++ src/main/java/exceptions/InvalidMarkException.java
@@ -1,0 +1,5 @@
+package exceptions;
+
+public class InvalidMarkException extends Exception {
+    // no further code
+}
