+++ src/main/java/exceptions/InvalidTodoException.java
@@ -1,0 +1,5 @@
+package exceptions;
+
+public class InvalidTodoException extends Exception {
+    // no further code
+}
