+++ src/main/java/exceptions/InvalidUnmarkException.java
@@ -1,0 +1,5 @@
+package exceptions;
+
+public class InvalidUnmarkException extends Exception {
+    // no further code
+}
