+++ src/main/java/exceptions/UnrecognisedCommandException.java
@@ -1,0 +1,5 @@
+package exceptions;
+
+public class UnrecognisedCommandException extends Exception{
+    // no further code
+}
