+++ src/main/java/tasks/Task.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package tasks;
 
 public class Task {
     public String descriptionString;
