+++ text-ui-test/EXPECTED.TXT
@@ -3,8 +3,6 @@ Hello! I'm Juke!
 What can I do for you?
 ------------------------
 
-	Out-of-bounds value entered, try again.
-
 	------------------------
 	Got it. I've added this task:
 	[T] [ ] first task
